Uploading store dashboard
The aim of this was to create a dashboard showcasing my findings after analysing a dataset that consisted of customer online shopping data. The dashboard and reports were made in Microsoft Excel. This mini project required several different steps such as data processing and analysing along with data visualization of the findings.